testbench: rename include folder
