+++ tools/testbench/include/testbench/trace.h
